2024/08/17 (junsung)
- 로컬스토리지에 id / nickname저장
- CreateBlog -> CreatePost로 변경
- 데이터 들어있는 변수 명 변경
- 불필요 코드 삭제
