--- b_c/src/Component/Blog/Function/Toolbar.js
@@ -1,89 +1,0 @@
-import React from "react";
-import Quill from "quill";
-import "quill/dist/quill.snow.css";
-
-const Size = Quill.import("attributors/style/size");
-Size.whitelist = [
-  "8px",
-  "10px",
-  "12px",
-  "14px",
-  "16px",
-  "18px",
-  "20px",
-  "22px",
-  "24px",
-  "26px",
-  "28px",
-  "32px",
-  "36px",
-  "40px",
-];
-Quill.register(Size, true);
-
-const Toolbar = React.forwardRef((props, ref) => {
-  return (
-    <div id="toolbar" ref={ref} className="toolbar">
-      <span className="ql-formats">
-        <select className="ql-header" defaultValue="">
-          <option value="1"></option>
-          <option value="2"></option>
-          <option value=""></option>
-        </select>
-        <select className="ql-font"></select>
-        <select className="ql-size" defaultValue="24px">
-          <option value="8px">8px</option>
-          <option value="10px">10px</option>
-          <option value="12px">12px</option>
-          <option value="14px">14px</option>
-          <option value="16px">16px</option>
-          <option value="18px">18px</option>
-          <option value="20px">20px</option>
-          <option value="22px">22px</option>
-          <option value="24px">24px</option>
-          <option value="26px">26px</option>
-          <option value="28px">28px</option>
-          <option value="32px">32px</option>
-          <option value="36px">36px</option>
-          <option value="40px">40px</option>
-        </select>
-      </span>
-      <span className="ql-formats">
-        <button className="ql-bold"></button>
-        <button className="ql-italic"></button>
-        <button className="ql-underline"></button>
-        <button className="ql-strike"></button>
-      </span>
-      <span className="ql-formats">
-        <select className="ql-color"></select>
-        <select className="ql-background"></select>
-      </span>
-      <span className="ql-formats">
-        <button className="ql-script" value="sub"></button>
-        <button className="ql-script" value="super"></button>
-      </span>
-      <span className="ql-formats">
-        <button className="ql-list" value="ordered"></button>
-        <button className="ql-list" value="bullet"></button>
-        <button className="ql-indent" value="-1"></button>
-        <button className="ql-indent" value="+1"></button>
-      </span>
-      <span className="ql-formats">
-        <select className="ql-align"></select>
-      </span>
-      <span className="ql-formats">
-        <button className="ql-direction" value="rtl"></button>
-      </span>
-      <span className="ql-formats">
-        <button className="ql-link"></button>
-        <button className="ql-image"></button>
-        <button className="ql-video"></button>
-      </span>
-      <span className="ql-formats">
-        <button className="ql-clean"></button>
-      </span>
-    </div>
-  );
-});
-
-export default Toolbar;
