+++ b_c/src/Component/Router.js
@@ -11,7 +11,7 @@ import ChangePW from "./ChangePW.js";
 import SelectID from "./SelectID.js";
 import ChangeUserInfo from "./ChangeUserInfo.js";
 import Blog from "./Blog/Blog.js";
-import CreateBlog from "./Blog/CreateBlog.js";
+import CreatePost from "./Blog/CreatePost.js";
 
 import TestModule from "./test_module (이후 삭제 예정)/TestModule.js"; // 테스트 목적의 데이터
 
@@ -27,7 +27,7 @@ function Router() {
       <Route path="/ChangePW" element={<ChangePW />} />
       <Route path="/HomePageLogin" element={<HomePageLogin />} />
       <Route path="/Blog" element={<Blog />} />
-      <Route path="/CreateBlog" element={<CreateBlog />} />
+      <Route path="/CreatePost" element={<CreatePost />} />
       <Route path="/Community" element={<Community />} />
       <Route path="/ChangeUserInfo" element={<ChangeUserInfo />} />
       <Route path="/TestModule" element={<TestModule />} />
