+++ b_c/src/Function/Menubar/MenubarLoginafter.js
@@ -45,7 +45,7 @@ function MenubarLoginafter() {
                   블로그
                 </h4>
                 <h5
-                  onClick={() => handleClick("/CreateBlog")}
+                  onClick={() => handleClick("/CreatePost")}
                   style={{ cursor: "pointer" }}
                 >
                   블로그 작성
